# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LA GLORIA STORE API",
        "description": "THE OFFICIAL LA GLORIA F.C. STORE",
        "contact": {
            "email": "dieztomasjoaquin@gmail.com"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/rest/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Display a listing of products",
                "description": "Get a paginated list of products with their details",
                "operationId": "a2e6f3dd87f8986f420c13f1b2244d16",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/rest/products/id/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Show a single product",
                "description": "Display the specified product.",
                "operationId": "showProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to show",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/rest/products/brand/{brandName}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get products filtering by brand name",
                "description": "Returns a list of products based on the provided brand name.",
                "operationId": "8d1e995bdd6f271d30de511b4afdcbdc",
                "parameters": [
                    {
                        "name": "brandName",
                        "in": "path",
                        "description": "The name of the brand",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The brand doesnt exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rest/products/category/{categoryName}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get products ",
                "description": "Returns a list of products based on the provided category name.",
                "operationId": "85106f6f2c5d19b3c4844b9025efe1ed",
                "parameters": [
                    {
                        "name": "categoryName",
                        "in": "path",
                        "description": "The name of the category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The category doesnt exist."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "BrandResource": {
                "title": "Brand resource",
                "description": "Represents a Brand",
                "properties": {
                    "id": {
                        "description": "The brand's ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "The brand's name",
                        "type": "string",
                        "example": "Nike"
                    },
                    "created_at": {
                        "description": "The datetime when the brand was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-05-11 13:42:56"
                    },
                    "updated_at": {
                        "description": "The datetime when the brand was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-05-11 13:42:56"
                    }
                },
                "type": "object"
            },
            "CategoryResource": {
                "title": "Category Resource",
                "description": "Represents a category",
                "properties": {
                    "id": {
                        "description": "The ID of the category",
                        "type": "integer",
                        "example": "2"
                    },
                    "name": {
                        "description": "The name of the category",
                        "type": "string",
                        "example": "Shoes"
                    },
                    "created_at": {
                        "description": "The date and time when the category was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The date and time when the category was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ProductResource": {
                "title": "ProductResource",
                "description": "Represents a Product",
                "properties": {
                    "id": {
                        "description": "The product's unique identifier",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the product",
                        "type": "string"
                    },
                    "size": {
                        "description": "The size of the product",
                        "type": "string"
                    },
                    "image": {
                        "description": "The URL of the product's image",
                        "type": "string"
                    },
                    "price": {
                        "description": "The price of the product",
                        "type": "number",
                        "format": "float"
                    },
                    "brand": {
                        "$ref": "#/components/schemas/BrandResource"
                    },
                    "category": {
                        "$ref": "#/components/schemas/CategoryResource"
                    },
                    "created_at": {
                        "description": "The date and time when the product was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The date and time when the product was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    }
}